Update script for ROC-curve and combined results of 3 networks
- Update existing validation script to plot ROC-curve for individual
molecules and for combined all molecules
- New script for specifically generate the results from combined
modelling results of the 3 networks using the rule of majority (without
zero) by combining rowSums and take the sign of sum results

Note1: Might still need to check carefully why the results seem to be
too good to be true for combined ones with thresholds.

Note2: Results are saved in separated folders to save committing time.

Former-commit-id: a736e68f8dbc25910e83a063668554574afe1fc4 [formerly eb5eca72764f950ebee31092aa45cb8524bb5645] [formerly 8de9ca42ec1ed5f34686d1a8a3cff84897acccff]
Former-commit-id: d070a9f8a9e1c2c6dc8a52d63bca2af1dfc40dfa
Former-commit-id: 20f3625c236be828f2883e7f0dd97f8e29729709