Fix issue with measurementsWeight mapping
- The indexing of absDiff starts from measurements first followed by
non-measured nodes e.g. 8 9 10 1 2 3 4 5 6 7 for Ex5. Therefore, the
mapping via expNodesReduced also have to change according to the
position of absDiff (stored in “idxMeasured”).

- Add measurementsWeights in Ex3 as a case study and obtained correct
results [!!NOTE!!] CHANGE OF THE CONCEPT - The measurements with higher
(DoRothEA) score -> have MORE weights (instead of less weights) ->
Therefore cplex try to fit the one with higher weights first to lower
the overall objective function — Also, the weights are now considered
as an “add-on” part instead of “soften” part in the objective function
so that it will not interfere with the influence from beta once the the
ratio alpha/beta is low.


Former-commit-id: 4e157afeeea71757f81698ebe49401d0054f725c
Former-commit-id: f143780c16b45d70ec1b10ee83e4e4a753922809
Former-commit-id: a43ee42e42058b676fd00787d1b51d56ae3a9de3