Revise CARNIVAL input format and clean up Uniprot2GeneSymbol
- According to the received feedbacks, specifying the input variables
directly in R would be easier and less error-prone for the users
comparing to the modifying a text file. Also, it would be easier for
the parallelisation on the cluster.

- Now the “runCARNIVAL” function takes all input arguments directly.
Descriptions for all variables are available in the help file. The QC
step for input format is also introduced to ensure that the user
plugged in the correct set of input information.

- The vignette of CARNIVAL is updated accordingly. An example to run
CARNIVAL with built-in in example and user-provided set of inputs are
available.

- Mod: clean up to Uniprot2GeneSymbol.R script to summarise and export
a unique set of unmapped node names to gene symbol instead of reporting
them on multiple lines.


Former-commit-id: 242113ef52779ec9c54758c3aef56c689bac9a9e [formerly 0f535d40dd0eaa4d3a714c7bd0353cfe3ac36b41] [formerly 9a711e66de6f2a3be520bfb23eb2e86ea68a793f] [formerly 921bfad962ddddc2a6eb5b8a57127d2950ba9c32]
Former-commit-id: dfc81b59c04258a85e326017c01035eb081d3980
Former-commit-id: 0369db47a68dadc11940678c0753ccee7aa7ea07
Former-commit-id: 147ff9f924343d8d4728150b38d30dfdda7252c2