Update runCARNIVAL.R
Issue: Upon running CARNIVAL on the cluster with a separated (driver) R-script, users couldn't directly pass the arguments from the cluster (argsJob) directly to the runCARNIVAL function.

Fix: Replaced the parameter "parallelCR" (T/F) with two new parameters "parallelIdx1" and "parallelIdx2" to allow a direct assignment of the indexing-number of intermediate and result files which should be unique on the cluster. Note that the user have to include the following lines on their driver script to allow parallelisation

library(doParallel)
argsJob=commandArgs(trailingOnly = TRUE)

Also, the runCARNIVAL function needs additional assignment for the new parameters:

runCARNIVAL(parallelIdx1 = as.numeric(argsJob[1]),
                              parallelIdx2 = as.numeric(argsJob[2]))

This new implementation will also be documented in the README and Wiki section upon merging.


Former-commit-id: b4b254ee3c26766e0dd5e2bbf90bd85c8f2b3c48 [formerly c97fbbfb1237cdb7176407fabba7611ee6329baf] [formerly 25b6039aeaada9384075db81eaac4e622fd176b5] [formerly 9796f43e116e09f3f1e2d9bbcdf522b6164b7ef6]
Former-commit-id: 4895847aa773de87b6e9eb5d15c1fde5da7c042e
Former-commit-id: 99dc39dc49062e1f429177fbb874c2ba69b2acc8
Former-commit-id: 42225f3df497c91e60dacd57e041b46423f85e84