Update runCARNIVAL.R
Issue: Upon running CARNIVAL on the cluster with a separated (driver) R-script, users couldn't directly pass the arguments from the cluster (argsJob) directly to the runCARNIVAL function.

Fix: Replaced the parameter "parallelCR" (T/F) with two new parameters "parallelIdx1" and "parallelIdx2" to allow a direct assignment of the indexing-number of intermediate and result files which should be unique on the cluster. Note that the user have to include the following lines on their driver script to allow parallelisation

library(doParallel)
argsJob=commandArgs(trailingOnly = TRUE)

Also, the runCARNIVAL function needs additional assignment for the new parameters:

runCARNIVAL(parallelIdx1 = as.numeric(argsJob[1]),
                              parallelIdx2 = as.numeric(argsJob[2]))

This new implementation will also be documented in the README and Wiki section upon merging.


Former-commit-id: 9796f43e116e09f3f1e2d9bbcdf522b6164b7ef6
Former-commit-id: 25b6039aeaada9384075db81eaac4e622fd176b5
Former-commit-id: c97fbbfb1237cdb7176407fabba7611ee6329baf