Update runDoRothEA.R
Fix the hidden bug when calling the object "viper_regulon" instead of "regulon" inside the function.

This function didn't return an error so far, possible due to the fact that the object "viper_regulon" was loaded into the workspace prior to calling the function. Nevertheless, it is a standard practice to correct all detected bugs to prevent future issues.


Former-commit-id: 977846fab3134aa24a97766e1ae2df112510df57
Former-commit-id: eb5705b99b3260eacb5ea09cb54cc32f938d4877
Former-commit-id: 488a9de97a8cc5ea2c1fa7e3b57170fcebc181d5