Update runDoRothEA.R
Fix the hidden bug when calling the object "viper_regulon" instead of "regulon" inside the function.

This function didn't return an error so far, possible due to the fact that the object "viper_regulon" was loaded into the workspace prior to calling the function. Nevertheless, it is a standard practice to correct all detected bugs to prevent future issues.


Former-commit-id: 427b4f7ba424a84cd411023b825988e5ec449f3a [formerly 488a9de97a8cc5ea2c1fa7e3b57170fcebc181d5] [formerly eb5705b99b3260eacb5ea09cb54cc32f938d4877] [formerly 977846fab3134aa24a97766e1ae2df112510df57]
Former-commit-id: 524692eb818d6d5561fd62e3292793bc0a8d64e3
Former-commit-id: 16fa843fdfb6149c267ad5cb52f7daccbc9c87d6
Former-commit-id: ee95f6a2f1dd98ab9ccd2b04c35830ce704999e8